refactor: apply constant file
