+++ apps/web/src/services/my-page/components/UserAccountMultiFactorAuthItems.vue
@@ -7,10 +7,11 @@ import {
 
 import { store } from '@/store';
 
-import { useMultiFactorAuthStore } from '@/services/my-page/stores/multi-factor-auth-store';
 import {
-    MULTI_FACTOR_AUTH_ITEMS, MULTI_FACTOR_AUTH_TYPE,
-} from '@/services/my-page/types/multi-factor-auth-type';
+    MULTI_FACTOR_AUTH_ITEMS,
+    MULTI_FACTOR_AUTH_TYPE,
+} from '@/services/my-page/constants/multi-factor-auth-constants';
+import { useMultiFactorAuthStore } from '@/services/my-page/stores/multi-factor-auth-store';
 
 const multiFactorAuthStore = useMultiFactorAuthStore();
 const multiFactorAuthState = multiFactorAuthStore.state;
